As suggested by the TODO list on GitHub, this commit extracts the message handling routines from the client classes and isolates them into specialized message handling classes. MessageHandler.java is the base class from which all other message handlers are derived.
PlayerMessageHandler.java is the message handler for clients which have
an associated Swing UI. It handles many of the game messages sent to the
client. If an unrecognized message is received it passes it on to the
super class for handling.

InProcMessageHandler.java is a message handler for Swing clients but which
is specialized for handling messages received from an in-process server.
Unhandled messages are passed to the PlayerMessageHandler class, and may
eventually be handled by the base MessageHandler super class.

RobotMessageHandler is the message handler for robot clients. It passes
unhandled messages to the base MessageHandler class. Much of the code
that handles messages destined for robot clients is still in RobotBrain.java.
It will be necessary to continue to extract that code into RobotMessageHandler.